Removed the old flac parsing system.
